% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing.R
\name{price_european_option}
\alias{price_european_option}
\title{Price European Option via Monte Carlo}
\usage{
price_european_option(price_paths, K, r, maturity, type = c("call", "put"))
}
\arguments{
\item{price_paths}{A matrix of simulated price paths (rows=steps, cols=paths).
The last row is assumed to be the price at maturity S_T.}

\item{K}{Strike price.}

\item{r}{Risk-free rate (annualized).}

\item{maturity}{Time to maturity (in years).}

\item{type}{Option type: "call" or "put".}
}
\value{
A list containing:
\item{price}{Estimated option price.}
\item{std_error}{Standard error of the Monte Carlo estimate.}
\item{conf_int}{95\\% confidence interval for the price.}
}
\description{
Calculates the price of a European option given simulated risk-neutral price paths.
}
