% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_gbm}
\alias{sim_gbm}
\title{Simulate Geometric Brownian Motion}
\usage{
sim_gbm(n, h, frequency = "daily", S0 = 100, mu = 0.05, sigma = 0.2, ...)
}
\arguments{
\item{n}{Number of time steps}

\item{h}{Horizon in years}

\item{frequency}{Frequency of simulation (e.g. "daily")}

\item{S0}{Initial price}

\item{mu}{Drift}

\item{sigma}{Volatility}

\item{...}{Additional arguments passed to esgtoolkit::simdiff}
}
\value{
A time series object of simulated paths
}
\description{
Simulate Geometric Brownian Motion
}
