% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_heston}
\alias{sim_heston}
\title{Simulate Heston Model}
\usage{
sim_heston(n, r0 = 0.05, ...)
}
\arguments{
\item{n}{Number of time steps}

\item{r0}{Initial return/drift parameter}

\item{...}{Additional arguments passed to esgtoolkit::rsvjd}
}
\value{
A time series object of simulated paths
}
\description{
Uses SVJD with lambda=0 (no jumps).
}
