% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_svjd}
\alias{sim_svjd}
\title{Simulate Stochastic Volatility with Jumps (SVJD)}
\usage{
sim_svjd(n, r0 = 0.05, ...)
}
\arguments{
\item{n}{Number of time steps}

\item{r0}{Initial return/drift parameter}

\item{...}{Additional arguments passed to esgtoolkit::rsvjd}
}
\value{
A time series object of simulated paths
}
\description{
Simulate Stochastic Volatility with Jumps (SVJD)
}
